chore(dashboard): bring the hrm alive
