+++ apps/dashboard/src/components/error-page.tsx
@@ -1,7 +1,7 @@
 import { useRouteError } from 'react-router-dom';
 
 export default function ErrorPage() {
-  const error = useRouteError() as any;
+  const error = useRouteError() as { statusText?: string; message: string };
   console.error(error);
 
   return (
