+++ apps/dashboard/src/context/auth/hooks.ts
@@ -1,0 +1,4 @@
+import { AuthContext } from './auth-context';
+import { createContextHook } from '@/utils/context';
+
+export const useAuth = createContextHook(AuthContext);
