+++ apps/dashboard/src/context/auth/index.ts
@@ -1,2 +1,3 @@
 export * from './auth-provider';
 export * from './auth-context';
+export * from './hooks';
