+++ apps/dashboard/src/context/environment/index.ts
@@ -1,1 +1,2 @@
 export * from './environment-provider';
+export * from './hooks';
