+++ apps/dashboard/src/context/index.ts
@@ -1,4 +1,4 @@
 export * from './clerk-provider';
-export * from './segment-provider';
+export * from './segment';
 export * from './auth';
 export * from './environment';
