+++ apps/dashboard/tests-examples/demo-todo-app.spec.ts
@@ -402,14 +402,14 @@ async function checkNumberOfTodosInLocalStorage(page: Page, expected: number) {
 
 async function checkNumberOfCompletedTodosInLocalStorage(page: Page, expected: number) {
   return await page.waitForFunction((e) => {
-    return JSON.parse(localStorage['react-todos']).filter((todo: any) => todo.completed).length === e;
+    return JSON.parse(localStorage['react-todos']).filter((todo) => todo.completed).length === e;
   }, expected);
 }
 
 async function checkTodosInLocalStorage(page: Page, title: string) {
   return await page.waitForFunction((t) => {
     return JSON.parse(localStorage['react-todos'])
-      .map((todo: any) => todo.title)
+      .map((todo) => todo.title)
       .includes(t);
   }, title);
 }
